Redesign report generation tools
